+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Model `fill()`
 - Model `getInstance()`
+- Model hooks `onFirstSaveHook()` and `onSaveHook()`
 - Section about Reusing models
 
 ### Changed